Refactor interfaces
what:
- Spread interfaces to into categories
