+++ ubc-scheduler/src/Definitions/Interfaces/CourseInterfaces.ts
@@ -16,18 +16,6 @@ export interface ICourseNumbers {
     numbers: number[]
 };
 
-export interface ISelectedData {
-    numeric?: number,
-    written?: string
-}
-
-export interface IQueryBuilderEntry {
-    courseCode: string,
-    year?: number,
-    courseNumber?: number,
-    term?: number 
-};
-
 export interface ISchedule {
     day: string,
     starttime: number,
