+++ ubc-scheduler/src/Definitions/Interfaces/QueryBuilderInterfaces.ts
@@ -1,0 +1,12 @@
+export interface ISelectedData {
+    numeric?: number,
+    written?: string
+}
+
+export interface IQueryBuilderEntry {
+    courseCode: string,
+    year?: number,
+    courseNumber?: number,
+    term?: number 
+};
+
