+++ ubc-scheduler/src/components/FilterSelection/SelectedCourseDisplay.tsx
@@ -3,7 +3,7 @@ import List from '@mui/material/List';
 import ListItem from '@mui/material/ListItem';
 import ListItemText from '@mui/material/ListItemText';
 import Divider from '@mui/material/Divider';
-import { IQueryBuilderEntry } from '../../Definitions/Interfaces/CommonInterfaces';
+import { IQueryBuilderEntry } from '../../Definitions/Interfaces/QueryBuilderInterfaces';
 import { selectQueriedCourses } from '../../Selectors/queryBuilderSelectors';
 import { shallowEqual, useSelector } from 'react-redux';
